+++ src/Heap.ts
@@ -6,38 +6,41 @@ export interface HeapOptions<T> extends VectorOptions<T> {
   compare?: (a: T, b: T) => boolean;
 }
 
-/**
- * @ignore
- */
 export class BinaryMinHeap<T> {
 
   public get size(): number {
-    return this._vector.size;
+    return this.vector.size;
   }
 
   constructor(
-      public _vector: Vector<T>
-    , public compare: (a: T, b: T) => boolean,
+    /**
+     * @ignore
+     */
+    public vector: Vector<T>,
+    /**
+     * @ignore
+     */
+    public compare: (a: T, b: T) => boolean,
   ) {
 
   }
 
   public min(): T {
-    if (this._vector.size === 0) {
+    if (this.vector.size === 0) {
       throw new Error(`Cannot get smallest element: heap is empty.`);
     }
-    return this._vector.first();
+    return this.vector.first();
   }
 
   public add(el: T): [boolean, Cursor<T>] {
-    this._vector.append(el);
-    const position = this._siftUp(this._vector.size);
-    return [true, new VectorCursor<T>(this._vector, position - 1)];
+    this.vector.append(el);
+    const position = this.siftUp(this.vector.size);
+    return [true, new VectorCursor<T>(this.vector, position - 1)];
   }
 
   public delete(el: T): boolean {
-    for (let i = 0; i < this._vector.size; ++i) {
-      if (this._vector.getAt(i) === el) {
+    for (let i = 0; i < this.vector.size; ++i) {
+      if (this.vector.getAt(i) === el) {
         this.deleteAtIndex(i);
         return true;
       }
@@ -47,8 +50,8 @@ export class BinaryMinHeap<T> {
 
   public deleteAll(el: T): number {
     let count = 0;
-    for (let i = 0; i < this._vector.size; ++i) {
-      if (this._vector.getAt(i) === el) {
+    for (let i = 0; i < this.vector.size; ++i) {
+      if (this.vector.getAt(i) === el) {
         this.deleteAtIndex(i);
         count++;
       }
@@ -61,20 +64,20 @@ export class BinaryMinHeap<T> {
   }
 
   public deleteAtIndex(index: number): void {
-    if (index === this._vector.size - 1) {
-      this._vector.deleteAtIndex(index);
+    if (index === this.vector.size - 1) {
+      this.vector.deleteAtIndex(index);
       return;
     }
-    const replacement = this._vector.getAt(this._vector.size - 1);
-    this._vector.replace(index, replacement);
-    this._vector.deleteAtIndex(this._vector.size - 1);
-    if (this._vector.size <= 1) {
+    const replacement = this.vector.getAt(this.vector.size - 1);
+    this.vector.replace(index, replacement);
+    this.vector.deleteAtIndex(this.vector.size - 1);
+    if (this.vector.size <= 1) {
       return;
     }
-    if (index > 1 && !this.compare(replacement, this._vector.getAt(Math.floor((index + 1) / 2) - 1))) {
-      this._siftUp(index + 1);
+    if (index > 1 && !this.compare(replacement, this.vector.getAt(Math.floor((index + 1) / 2) - 1))) {
+      this.siftUp(index + 1);
     } else {
-      this._siftDown(index + 1);
+      this.siftDown(index + 1);
     }
   }
 
@@ -83,21 +86,21 @@ export class BinaryMinHeap<T> {
   }
 
   public clone(): BinaryMinHeap<T> {
-    return new BinaryMinHeap<T>(this._vector.clone(), this.compare);
+    return new BinaryMinHeap<T>(this.vector.clone(), this.compare);
   }
 
-  private _siftDown(position: number) {
+  private siftDown(position: number) {
     while (true) {
       const left = 2 * position, right = 2 * position + 1;
       let smallest = position;
-      if (left <= this._vector.size && this.compare(this._vector.getAt(left - 1), this._vector.getAt(smallest - 1))) {
+      if (left <= this.vector.size && this.compare(this.vector.getAt(left - 1), this.vector.getAt(smallest - 1))) {
         smallest = left;
       }
-      if (right <= this._vector.size && this.compare(this._vector.getAt(right - 1), this._vector.getAt(smallest - 1))) {
+      if (right <= this.vector.size && this.compare(this.vector.getAt(right - 1), this.vector.getAt(smallest - 1))) {
         smallest = right;
       }
       if (smallest !== position) {
-        this._vector.swap(position - 1, smallest - 1);
+        this.vector.swap(position - 1, smallest - 1);
         position = smallest;
       } else {
         break;
@@ -105,15 +108,15 @@ export class BinaryMinHeap<T> {
     }
   }
 
-  private _siftUp(position: number) {
-    let element = this._vector.getAt(position - 1);
+  private siftUp(position: number) {
+    let element = this.vector.getAt(position - 1);
     while (position > 1) {
       const parentPos = Math.floor(position / 2);
-      if (!this.compare(this._vector.getAt(parentPos - 1), this._vector.getAt(position - 1))) {
+      if (!this.compare(this.vector.getAt(parentPos - 1), this.vector.getAt(position - 1))) {
         // swap() inlined to save one getAt() call
-        const keep = this._vector.getAt(parentPos - 1);
-        this._vector.replace(parentPos - 1, element);
-        this._vector.replace(this._vector.size - 1, keep);
+        const keep = this.vector.getAt(parentPos - 1);
+        this.vector.replace(parentPos - 1, element);
+        this.vector.replace(this.vector.size - 1, keep);
         element = keep;
         position = parentPos;
       } else {